♻️ Restructure method detail components and improve code organization
- Move ResponseTabs, CodeExample, and SyntaxHighlighter to dedicated subdirectories
- Create centralized components index for easier imports
- Enhance type handling and error checking in components
- Improve styling and responsiveness of method detail sections
