+++ packages/frontend/src/components/method-detail/components/code/CodeExample.tsx
