+++ packages/frontend/src/components/method-detail/components/code/SyntaxHighlighter.tsx
