+++ packages/frontend/src/components/method-detail/components/index.ts
@@ -1,0 +1,8 @@
+export * from './code/CodeExample';
+export * from './code/SyntaxHighlighter';
+export * from './tabs/ResponseTabs';
+export * from './sections/ErrorsSection';
+export * from './sections/ExamplesSection';
+export * from './sections/MethodHeader';
+export * from './sections/ParametersSection';
+export * from './sections/ResultSection';
