+++ packages/frontend/src/components/method-detail/components/sections/MethodHeader.tsx
