+++ packages/frontend/src/components/method-detail/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Layout';
+export * from './components';
